Хер знает, как теперь писать коммит... Дохрена чего сделал, надо было делать промежуточные коммиты, пропробую описать все в одном коммите и ничего не упустить 1. Решил делать отсчет длительности поездки в игровых часах (всплывающее окно Road в игре изменил соответсвующим образом) 2. Изменил твой скрипт времени Timeflow и GameTime, изменения небольшие, чисто для удобства кое-что сделал и удалил флаг TimeFlow он там нахер не нужен был 3. Теперь переход по кнопке "В путь" переносит на сцену поездки 4. Добавил счетчик оставшегося времени пути (на канвасе TravelBlock (охереть TortoiseGit помогает названия искать(даже объектов на сцене, я в ахере))) 5. Еще всякой фигни по мелоче, сейчас даже не вспомню что там

+++ Assets/KirillTMP/Road.cs
@@ -12,8 +12,8 @@ public class Road : MonoBehaviour
     [SerializeField] private string _description;
     public string Description => _description;
     
-    [SerializeField] private float _travelingTime; // пока неясно в каких единицах будет измеряться (игровых часах
-    public float TravelingTime => _travelingTime;  // или в реальных секундах/минутах) 
+    [SerializeField] private int _travelingTime; // измеряется в игровых часах
+    public int TravelingTime => _travelingTime;  
     [SerializeField] private int _quality;
     public int Quality => _quality;
     [SerializeField] private int _danger;
